Create function to perform gradient descent
The function is used to implement the gradient descent algorithm given a
number of parameters which are:
 - A function to calculate the Slope (Derivative) of the Cost Function.
 - A value of x to start the gradient descent algorithm.
 - A multiplier which is used to control the step size.
 - A precision which is used to terminate the for loop when the change
   in the cost function is less than this value.

Note: The x_list and slope_list lists are initialised to empty lists.
The tutorial actually initialises to contain a value. This changes the
length of the lists returned, and hence the number of steps the
algorithm actually takes.